shopping service
Former-commit-id: 1d5feecb8afeec1dfdc4826ff75bf2e25d587b76

+++ core/domain/interface/merchant/merchant.go
@@ -129,7 +129,7 @@ type (
 		CreateSignUpToken(memberId int64) string
 
 		// 根据商户申请密钥获取会员编号
-		GetMemberFromSignUpToken(token string) int32
+		GetMemberFromSignUpToken(token string) int64
 
 		// 提交商户注册信息
 		CommitSignUpInfo(*MchSignUp) (int32, error)
