+++ core/repository/merchant_repo.go
@@ -92,11 +92,11 @@ func (m *merchantRepo) CreateSignUpToken(memberId int64) string {
 }
 
 // 根据商户申请密钥获取会员编号
-func (m *merchantRepo) GetMemberFromSignUpToken(token string) int32 {
+func (m *merchantRepo) GetMemberFromSignUpToken(token string) int64 {
 	key := "go2o:rep:mch:signup:tk-" + token
-	id, err := m.storage.GetInt(key)
+	id, err := m.storage.GetInt64(key)
 	if err == nil {
-		return int32(id)
+		return id
 	}
 	return -1
 }
